Editorial: provide a shorthand for other specifications to create events
This can be useful when you can’t use fire, but need to initialize
event attributes in a particular manner or keep the event around for
some reason.

Note that creating events without a constructor doesn’t work, as you
would not get the dictionary to usefully default attribute values.